Renamed files and reorganized functions
