+++ src/analysis/mod.rs
@@ -1,1 +1,3 @@
+//! Analysis Tools.
+
 pub mod validator;
