+++ src/analysis/validator.rs
@@ -1,26 +1,41 @@
-extern crate llvm_sys as llvm;
+//! This module provides functionality for checking the correctness of LLVM modules and functions.
 
-use std::sync::{Arc, RwLock};
+extern crate llvm_sys as llvm; 
+use std::sync::{Arc, RwLock}; 
+use llvm::{analysis, core}; 
+use crate::memory_management::pointer::{SafeLLVMPointer, LLVMRef, LLVMRefType}; 
 
-use llvm::{analysis, core};
-
-use crate::memory_management::pointer::{CPointer, LLVMRef, LLVMRefType};
-
+/// A Validator struct that encapsulates an LLVM module within a thread-safe, reference-counted pointer.
 pub struct Validator {
-    module: Arc<RwLock<CPointer>>,
+    module: Arc<RwLock<SafeLLVMPointer>>, // Encapsulated LLVM module pointer.
 }
 
 impl Validator {
-    // Constructs a new Validator
-    pub fn new(module: Arc<RwLock<CPointer>>) -> Self {
+    /// Constructs a new `Validator`.
+    ///
+    /// # Parameters
+    /// * `module` - An `Arc<RwLock<SafeLLVMPointer>>` pointing to the LLVM module to be checkd.
+    ///
+    /// # Returns
+    /// A new instance of `Validator`.
+    pub fn new(module: Arc<RwLock<SafeLLVMPointer>>) -> Self {
         Self { module }
     }
 
-    pub fn get_module(&self) -> &Arc<RwLock<CPointer>> {
+    /// Retrieves a shared reference to the encapsulated module.
+    ///
+    /// # Returns
+    /// A shared reference to the Arc<RwLock<SafeLLVMPointer>> of the module.
+    pub fn get_module(&self) -> &Arc<RwLock<SafeLLVMPointer>> {
        &self.module
     }
 
-    // Validates an entire LLVM module
+    /// Checks the entire LLVM module for correctness.
+    ///
+    /// This function uses LLVM's verification function to check the module and prints an error message if validation fails.
+    ///
+    /// # Returns
+    /// True if the module is valid, false otherwise.
     pub fn is_valid_module(&self) -> bool {
         let mut error_message = std::ptr::null_mut();
 
@@ -49,8 +64,16 @@ impl Validator {
         }
     }
 
-    // Validates a specific function in the module
-    pub fn is_valid_function(&self, function: Arc<RwLock<CPointer>>) -> bool {
+    /// Checks a specific function within the module for correctness.
+    ///
+    /// This function reads the function from its encapsulated pointer and Checks it using LLVM's function verification API.
+    ///
+    /// # Parameters
+    /// * `function` - An Arc<RwLock<SafeLLVMPointer>> pointing to the LLVM function to be checkd.
+    ///
+    /// # Returns
+    /// True if the function is valid, false otherwise.
+    pub fn is_valid_function(&self, function: Arc<RwLock<SafeLLVMPointer>>) -> bool {
         let action = analysis::LLVMVerifierFailureAction::LLVMPrintMessageAction;
 
         let function_rw_lock = function.read().expect("Failed to get function");
@@ -70,6 +93,5 @@ impl Validator {
             eprintln!("Function validation failed.");
             false
         }
-
     }
 }
