+++ src/common/cstring.rs
