+++ src/ir/mod.rs
@@ -5,7 +5,7 @@ pub mod builder;
 pub mod values;
 
 /// Create operations
-pub mod ops;
+pub mod binary_operations;
 
 /// Create a type
 pub mod types;
@@ -19,8 +19,4 @@ pub mod block;
 /// TLE related utilities
 pub mod top_level_exp;
 
-pub mod init;
-
-pub mod router;
-
-pub use router::SafeLLVM;
+pub mod core;
