+++ src/lib.rs
@@ -12,31 +12,12 @@ pub mod ir;
 /// Offers tools for Just-In-Time (JIT) compilation.
 pub mod jit;
 
-/// Manages handling LLVM objects safely.
-pub mod memory_management;
-
 /// Logging utilities to support internal debugging.
 pub mod logger;
 
-/// Helper utilities for common operations throughout the library.
-pub mod utils;
-
-/// Interface for external modules or libraries to interact with this project's functionalities.
-pub mod interface;
-
-// pub mod lld;  // TODO support for the LLVM LLD linker.
-
-// pub mod serialization;  // TODO serialization support of LLVM data structures.
-
-// pub mod debug;  // TODO debugging utilities for internal and external diagnostics.
-
-// pub mod aot;  // TODO tools related to Ahead-Of-Time compilation capabilities.
+pub mod common;
 
 /// Constants 
 pub mod constants;
 
 pub mod analysis; 
-
-pub mod router;
-
-pub use router::SafeLLVM;  // Exports the public api.
