+++ src/logger/mod.rs
@@ -1,3 +1,1 @@
-pub mod init;
-
-pub mod log;
+pub mod core;
