--- src/memory_management/mod.rs
@@ -1,7 +1,0 @@
-/// Thread safe pointer type for wrapping pointers.
-pub mod pointer;
-
-/// Resource manager used for IR generation.
-pub mod resource_pools;
-
-pub mod definitions;
