--- src/utils/utils_struct.rs
@@ -1,1 +1,0 @@
-pub struct Utils {}
