+++ tests/ir_tests/var_tests.rs
@@ -1,8 +1,8 @@
-use safe_llvm::memory_management::resource_pools::ResourcePools;
+use safe_llvm::memory_management::resource_pools::IRGenerator;
 
 #[test]
 fn test_init_var() {
-    let mut resource_pools = ResourcePools::new();
+    let mut resource_pools = IRGenerator::new();
 
     let context_tag = resource_pools.create_context().expect("Failed to create context");
     let module_tag = resource_pools.create_module("MyModule", context_tag).expect("Failed to create module");
@@ -23,7 +23,7 @@ fn test_init_var() {
 
 #[test]
 fn test_reassign_var() {
-    let mut resource_pools = ResourcePools::new();
+    let mut resource_pools = IRGenerator::new();
 
     let context_tag = resource_pools.create_context().expect("Failed to create context");
     let module_tag = resource_pools.create_module("MyModule", context_tag).expect("Failed to create module");
@@ -46,7 +46,7 @@ fn test_reassign_var() {
 
 #[test]
 fn test_get_var() {
-    let mut resource_pools = ResourcePools::new();
+    let mut resource_pools = IRGenerator::new();
 
     let context_tag = resource_pools.create_context().expect("Failed to create context");
     let module_tag = resource_pools.create_module("MyModule", context_tag).expect("Failed to create module");
