+++ tests/jit_tests/targeting_tests.rs
@@ -1,10 +1,10 @@
 use safe_llvm::{
-    analysis::validator::Validator, constants::{DEFAULT_BASIC_BLOCK_NAME, DEFAULT_FUNCTION_NAME, DEFAULT_MODULE_NAME}, jit::{execution_engine::ExecutionEngine, target::*}, memory_management::resource_pools::ResourcePools, utils::utils_struct::Utils
+    analysis::validator::Validator, constants::{DEFAULT_BASIC_BLOCK_NAME, DEFAULT_FUNCTION_NAME, DEFAULT_MODULE_NAME}, jit::{execution_engine::ExecutionEngine, target::*}, memory_management::resource_pools::IRGenerator, utils
 };
 
 #[test]
 fn test_execution_engine_with_general_targeting() {
-    let mut resource_pools = ResourcePools::new();
+    let mut resource_pools = IRGenerator::new();
 
     let context_tag = resource_pools.create_context().expect("Failed to create context");
     let module_tag = resource_pools.create_module(DEFAULT_MODULE_NAME, context_tag).expect("Failed to create module");
@@ -20,7 +20,7 @@ fn test_execution_engine_with_general_targeting() {
 
     let module = resource_pools.get_module(module_tag).expect("Failed to retrieve module");
 
-    match Utils::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
+    match utils::write_ir::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
         Ok(_) => {}
         Err(e) => {
             eprintln!("File write error: {}", e);
@@ -45,7 +45,7 @@ fn test_execution_engine_with_general_targeting() {
 
 #[test]
 fn test_execution_engine_with_arm_targeting() {
-    let mut resource_pools = ResourcePools::new();
+    let mut resource_pools = IRGenerator::new();
 
     let context_tag = resource_pools.create_context().expect("Failed to create context");
     let module_tag = resource_pools.create_module(DEFAULT_MODULE_NAME, context_tag).expect("Failed to create module");
@@ -61,7 +61,7 @@ fn test_execution_engine_with_arm_targeting() {
 
     let module = resource_pools.get_module(module_tag).expect("Failed to retrieve module");
 
-    match Utils::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
+    match utils::write_ir::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
         Ok(_) => {}
         Err(e) => {
             eprintln!("File write error: {}", e);
@@ -86,7 +86,7 @@ fn test_execution_engine_with_arm_targeting() {
 
 #[test]
 fn test_execution_engine_with_x86_targeting() {
-    let mut resource_pools = ResourcePools::new();
+    let mut resource_pools = IRGenerator::new();
 
     let context_tag = resource_pools.create_context().expect("Failed to create context");
     let module_tag = resource_pools.create_module(DEFAULT_MODULE_NAME, context_tag).expect("Failed to create module");
@@ -102,7 +102,7 @@ fn test_execution_engine_with_x86_targeting() {
 
     let module = resource_pools.get_module(module_tag).expect("Failed to retrieve module");
 
-    match Utils::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
+    match utils::write_ir::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
         Ok(_) => {}
         Err(e) => {
             eprintln!("File write error: {}", e);
@@ -127,7 +127,7 @@ fn test_execution_engine_with_x86_targeting() {
 
 #[test]
 fn test_execution_engine_with_mips_targeting() {
-    let mut resource_pools = ResourcePools::new();
+    let mut resource_pools = IRGenerator::new();
 
     let context_tag = resource_pools.create_context().expect("Failed to create context");
     let module_tag = resource_pools.create_module(DEFAULT_MODULE_NAME, context_tag).expect("Failed to create module");
@@ -143,7 +143,7 @@ fn test_execution_engine_with_mips_targeting() {
 
     let module = resource_pools.get_module(module_tag).expect("Failed to retrieve module");
 
-    match Utils::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
+    match utils::write_ir::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
         Ok(_) => {}
         Err(e) => {
             eprintln!("File write error: {}", e);
@@ -168,7 +168,7 @@ fn test_execution_engine_with_mips_targeting() {
 
 #[test]
 fn test_execution_engine_with_rv_targeting() {
-    let mut resource_pools = ResourcePools::new();
+    let mut resource_pools = IRGenerator::new();
 
     let context_tag = resource_pools.create_context().expect("Failed to create context");
     let module_tag = resource_pools.create_module(DEFAULT_MODULE_NAME, context_tag).expect("Failed to create module");
@@ -184,7 +184,7 @@ fn test_execution_engine_with_rv_targeting() {
 
     let module = resource_pools.get_module(module_tag).expect("Failed to retrieve module");
 
-    match Utils::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
+    match utils::write_ir::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
         Ok(_) => {}
         Err(e) => {
             eprintln!("File write error: {}", e);
@@ -209,7 +209,7 @@ fn test_execution_engine_with_rv_targeting() {
 
 #[test]
 fn test_execution_engine_with_wasm_targeting() {
-    let mut resource_pools = ResourcePools::new();
+    let mut resource_pools = IRGenerator::new();
 
     let context_tag = resource_pools.create_context().expect("Failed to create context");
     let module_tag = resource_pools.create_module(DEFAULT_MODULE_NAME, context_tag).expect("Failed to create module");
@@ -225,7 +225,7 @@ fn test_execution_engine_with_wasm_targeting() {
 
     let module = resource_pools.get_module(module_tag).expect("Failed to retrieve module");
 
-    match Utils::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
+    match utils::write_ir::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
         Ok(_) => {}
         Err(e) => {
             eprintln!("File write error: {}", e);
@@ -250,7 +250,7 @@ fn test_execution_engine_with_wasm_targeting() {
 
 #[test]
 fn test_execution_engine_with_ppc_targeting() {
-    let mut resource_pools = ResourcePools::new();
+    let mut resource_pools = IRGenerator::new();
 
     let context_tag = resource_pools.create_context().expect("Failed to create context");
     let module_tag = resource_pools.create_module(DEFAULT_MODULE_NAME, context_tag).expect("Failed to create module");
@@ -266,7 +266,7 @@ fn test_execution_engine_with_ppc_targeting() {
 
     let module = resource_pools.get_module(module_tag).expect("Failed to retrieve module");
 
-    match Utils::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
+    match utils::write_ir::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
         Ok(_) => {}
         Err(e) => {
             eprintln!("File write error: {}", e);
@@ -291,7 +291,7 @@ fn test_execution_engine_with_ppc_targeting() {
 
 #[test]
 fn test_execution_engine_with_sparc_targeting() {
-    let mut resource_pools = ResourcePools::new();
+    let mut resource_pools = IRGenerator::new();
 
     let context_tag = resource_pools.create_context().expect("Failed to create context");
     let module_tag = resource_pools.create_module(DEFAULT_MODULE_NAME, context_tag).expect("Failed to create module");
@@ -307,7 +307,7 @@ fn test_execution_engine_with_sparc_targeting() {
 
     let module = resource_pools.get_module(module_tag).expect("Failed to retrieve module");
 
-    match Utils::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
+    match utils::write_ir::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
         Ok(_) => {}
         Err(e) => {
             eprintln!("File write error: {}", e);
@@ -332,7 +332,7 @@ fn test_execution_engine_with_sparc_targeting() {
 
 #[test]
 fn test_execution_engine_with_systemz_targeting() {
-    let mut resource_pools = ResourcePools::new();
+    let mut resource_pools = IRGenerator::new();
 
     let context_tag = resource_pools.create_context().expect("Failed to create context");
     let module_tag = resource_pools.create_module(DEFAULT_MODULE_NAME, context_tag).expect("Failed to create module");
@@ -348,7 +348,7 @@ fn test_execution_engine_with_systemz_targeting() {
 
     let module = resource_pools.get_module(module_tag).expect("Failed to retrieve module");
 
-    match Utils::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
+    match utils::write_ir::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
         Ok(_) => {}
         Err(e) => {
             eprintln!("File write error: {}", e);
@@ -373,7 +373,7 @@ fn test_execution_engine_with_systemz_targeting() {
 
 #[test]
 fn test_execution_engine_with_aarch64_targeting() {
-    let mut resource_pools = ResourcePools::new();
+    let mut resource_pools = IRGenerator::new();
 
     let context_tag = resource_pools.create_context().expect("Failed to create context");
     let module_tag = resource_pools.create_module(DEFAULT_MODULE_NAME, context_tag).expect("Failed to create module");
@@ -389,7 +389,7 @@ fn test_execution_engine_with_aarch64_targeting() {
 
     let module = resource_pools.get_module(module_tag).expect("Failed to retrieve module");
 
-    match Utils::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
+    match utils::write_ir::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
         Ok(_) => {}
         Err(e) => {
             eprintln!("File write error: {}", e);
@@ -414,7 +414,7 @@ fn test_execution_engine_with_aarch64_targeting() {
 
 #[test]
 fn test_execution_engine_with_amdgpu_targeting() {
-    let mut resource_pools = ResourcePools::new();
+    let mut resource_pools = IRGenerator::new();
 
     let context_tag = resource_pools.create_context().expect("Failed to create context");
     let module_tag = resource_pools.create_module(DEFAULT_MODULE_NAME, context_tag).expect("Failed to create module");
@@ -430,7 +430,7 @@ fn test_execution_engine_with_amdgpu_targeting() {
 
     let module = resource_pools.get_module(module_tag).expect("Failed to retrieve module");
 
-    match Utils::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
+    match utils::write_ir::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
         Ok(_) => {}
         Err(e) => {
             eprintln!("File write error: {}", e);
@@ -455,7 +455,7 @@ fn test_execution_engine_with_amdgpu_targeting() {
 
 #[test]
 fn test_execution_engine_with_bpf_targeting() {
-    let mut resource_pools = ResourcePools::new();
+    let mut resource_pools = IRGenerator::new();
 
     let context_tag = resource_pools.create_context().expect("Failed to create context");
     let module_tag = resource_pools.create_module(DEFAULT_MODULE_NAME, context_tag).expect("Failed to create module");
@@ -471,7 +471,7 @@ fn test_execution_engine_with_bpf_targeting() {
 
     let module = resource_pools.get_module(module_tag).expect("Failed to retrieve module");
 
-    match Utils::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
+    match utils::write_ir::write_to_file(module.clone(), "test_execution_engine_with_general_targeting") {
         Ok(_) => {}
         Err(e) => {
             eprintln!("File write error: {}", e);
