--- tests/mock_tests/mod.rs
@@ -1,8 +1,0 @@
-// extern crate llvm_sys as llvm;
-
-// use mockall::automock;
-
-// use crate::interface::LLVMApi;
-
-// #[automock]
-// impl LLVMApi {}
