+++ tests/mod.rs
@@ -1,7 +1,3 @@
 pub mod ir_tests;
 
-pub mod memory_management_tests;
-
-pub mod mock_tests;
-
 pub mod jit_tests;
